abstract command nodes, also stopped the command thing from being bad
